requirement.txt
requests
pandas
numpy
scikit-learn
flask
python-dotenv
pandas-ta